amending upon comments: tidy up folders, remove and rename some  properties
