+++ backend/src/repositories/trainersService.ts
@@ -1,4 +1,5 @@
-import { Trainers, Trainer } from '../models/trainerModel.js';
+import { RowDataPacket } from 'mysql2';
+import { Trainers, Trainer } from '../types/trainer.js';
 import { makeDb } from '../db/db.js';
 
 // TODO : add pagination, handle errors
@@ -6,13 +7,10 @@ export const getTrainers = async (): Promise<Trainers> => {
   const db = await makeDb();
   const [rows] = await db.query(`
     SELECT 
-      p.id AS trainer_id,
+      p.id AS id,
       p.given_name,
       p.surname,
-      p.email,
       p.phone,
-      p.email_verified,
-      p.created_at,
       t.city,
       GROUP_CONCAT(DISTINCT d.name ORDER BY d.name SEPARATOR ', ') AS disciplines,
       GROUP_CONCAT(DISTINCT l.name ORDER BY l.name SEPARATOR ', ') AS languages
@@ -22,7 +20,7 @@ export const getTrainers = async (): Promise<Trainers> => {
     LEFT JOIN discipline d ON td.discipline_id = d.id
     LEFT JOIN trainer_lang tl ON t.person_id = tl.trainer_id
     LEFT JOIN lang l ON tl.lang_id = l.id
-    GROUP BY p.id, p.given_name, p.surname, p.email, p.phone, p.email_verified, p.created_at, t.city
+    GROUP BY p.id, p.given_name, p.surname, p.phone, t.city
   `);
 
   return rows as Trainers;
@@ -30,10 +28,10 @@ export const getTrainers = async (): Promise<Trainers> => {
 
 export const getTrainerById = async (id: string): Promise<Trainer> => {
   const db = await makeDb();
-  const [rows] = await db.query(
+  const [rows] = await db.query<Array<Trainer & RowDataPacket>>(
     `
     SELECT 
-      p.id AS trainer_id,
+      p.id AS id,
       p.given_name,
       p.surname,
       p.email,
@@ -50,12 +48,10 @@ export const getTrainerById = async (id: string): Promise<Trainer> => {
     LEFT JOIN trainer_lang tl ON t.person_id = tl.trainer_id
     LEFT JOIN lang l ON tl.lang_id = l.id
     WHERE p.id = ?
-    GROUP BY p.id, p.given_name, p.surname, p.email, p.phone, p.email_verified, p.created_at, t.city
+    GROUP BY p.id, p.given_name, p.surname, p.phone, p.created_at, t.city
     `,
     [id]
   );
 
-  // TODO : remove the any type
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return (rows as any[])[0] as Trainer;
+  return rows[0] as Trainer;
 };
